+++ llama/llava.py
@@ -4,6 +4,7 @@ __all__ = [
     'llava_completions',
 ]
 
+import gc
 import os
 from typing import Any, Optional, Iterator
 
@@ -98,7 +99,15 @@ def llava_completions(model: 'Model', model_options: ModelOptions, completions_o
     else:
         lib.llama_log_set(lib.llama_cpp_cffi_ggml_log_callback, ffi.NULL)
 
-    context = context_init(_model, model_options)
+    # context = context_init(_model, model_options)
+    try:
+        context = context_init(_model, model_options)
+    except MemoryError as e:
+        print('MemoryError:', e)
+        context = None
+        gc.collect()
+        return
+
     clip_context = clip_init_context(model_options)
     sampler = sampler_init(_model, completions_options)
     # print(f'{sampler=}')